requirements.txt
streamlit
python-dotenv
requests
groq
edge-tts
langchain
langchain-groq
langchain-tavily
langgraph
huggingface-hub
SpeechRecognition 
pydub
gTTS